fix: '[FE] 로그인시 세션 스토리지에 로그인 flag 저장'
이 플래그로 TopRoute에서 로컬스토리지를 초기화 할 지 확인합니다

+++ frontend/src/TopRoutes.tsx
@@ -33,7 +33,7 @@ const Router = () => {
   // 25분에 한번 accesstoken을 재발급받아온다
   useEffect(() => {
     if (flag) {
-      localStorage.clear();
+      if (!sessionStorage.getItem('isLogin')) localStorage.clear();
     }
     setFlag(!flag);
     const timer = setInterval(() => {
